requirements.txt
django
requests
beautifulsoup4
openai
python-dotenv
